+++ src/themes/createShadow.ts
@@ -40,13 +40,13 @@ export interface CreatedShadowTheme {
 const defaultOutline = `0 0 0 2px white, 0 0 0 4px ${defaultColor.OUTLINE}`
 
 export const defaultShadow = {
-  BASE: 'rgba(51, 51, 51, 0.15) 0 0 4px 0',
-  DIALOG: 'rgba(51, 51, 51, 0.3) 0 4px 10px 0',
+  BASE: `${defaultColor.TRANSPARENCY_15} 0 0 4px 0`,
+  DIALOG: `${defaultColor.TRANSPARENCY_30} 0 4px 10px 0`,
   LAYER0: 'none',
-  LAYER1: '0 1px 2px 0 rgba(0,0,0,0.24)',
-  LAYER2: '0 2px 4px 1px rgba(0,0,0,0.24)',
-  LAYER3: '0 4px 8px 2px rgba(0,0,0,0.24)',
-  LAYER4: '0 8px 16px 4px rgba(0,0,0,0.24)',
+  LAYER1: `0 1px 2px 0 ${defaultColor.TRANSPARENCY_30}`,
+  LAYER2: `0 2px 4px 1px ${defaultColor.TRANSPARENCY_30}`,
+  LAYER3: `0 4px 8px 2px ${defaultColor.TRANSPARENCY_30}`,
+  LAYER4: `0 8px 16px 4px ${defaultColor.TRANSPARENCY_30}`,
   OUTLINE: defaultOutline,
 }